move FAB to bottom navigation
