+++ app/src/main/res/layout/fragment_home.xml
@@ -23,18 +23,4 @@
 
     </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
 
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/fabCompose"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="25dp"
-        android:layout_marginEnd="25dp"
-        android:backgroundTint="@color/baby_blue"
-        android:clickable="true"
-        android:onClick="showPopup"
-        android:src="@drawable/ic_filled_compose"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:tint="@color/dark_blue" />
-
 </androidx.constraintlayout.widget.ConstraintLayout>
